backend-python/pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.115.13",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
]
